Adds unmodified xdrfile-1.1.4

+++ fortran/lib/xdrfile-1.1.4-d/src/xdrfile_trr.c
@@ -2,13 +2,33 @@
  *
  * $Id$
  *
- * Copyright (c) Erik Lindahl, David van der Spoel 2003,2004.
- * Coordinate compression (c) by Frans van Hoesel. 
+ /* -*- mode: c; tab-width: 4; indent-tabs-mode: t; c-basic-offset: 4 -*-
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public License
- * as published by the Free Software Foundation; either version 3
- * of the License, or (at your option) any later version.
+ * $Id$
+ *
+ * Copyright (c) 2009-2014, Erik Lindahl & David van der Spoel
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
 #include <stdlib.h>
@@ -404,7 +424,7 @@ static int do_htrn(XDRFILE *xd,mybool bRead,t_trnheader *sh,
 	return exdrOK;
 }
 
-static int do_trn(XDRFILE *xd,mybool bRead,int *step,double *t,double *lambda,
+static int do_trn(XDRFILE *xd,mybool bRead,int *step,float *t,float *lambda,
 				  matrix box,int *natoms,rvec *x,rvec *v,rvec *f)
 {
     t_trnheader *sh;
@@ -464,13 +484,13 @@ int read_trr_natoms(char *fn,int *natoms)
 	return exdrOK;
 }
 
-int write_trr(XDRFILE *xd,int natoms,int step,double t,double lambda,
+int write_trr(XDRFILE *xd,int natoms,int step,float t,float lambda,
 			  matrix box,rvec *x,rvec *v,rvec *f)
 {
 	return do_trn(xd,0,&step,&t,&lambda,box,&natoms,x,v,f);
 }
 
-int read_trr(XDRFILE *xd,int natoms,int *step,double *t,double *lambda,
+int read_trr(XDRFILE *xd,int natoms,int *step,float *t,float *lambda,
 			 matrix box,rvec *x,rvec *v,rvec *f)
 {
 	return do_trn(xd,1,step,t,lambda,box,&natoms,x,v,f);
